backend_api/requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
aspose-words==22.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.1.0
boto3==1.24.9
botocore==1.27.9
cairocffi==1.3.0
CairoSVG==2.5.2
celery==5.2.6
celery-progress==0.1.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect2==0.6.0
datefinder==0.7.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-anysign==1.4
django-axes==5.39.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-docusign==3.3
django-ipware==4.0.2
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-pgcrypto==2.0.0
django-seed==0.3.1
django-storages==1.12.3
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.1.0
django_filter==22.1
docusign-esign==3.16.0
drf-yasg==1.20.0
ecdsa==0.17.0
et-xmlfile==1.1.0
Faker==13.14.0
future==0.18.2
greenlet==1.1.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonfield==3.1.0
kombu==5.2.4
lxml==4.8.0
MarkupSafe==2.1.0
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.22.4
oauth2==1.9.0.post1
opencv-python==4.6.0.66
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pathspec==0.9.0
pdf2image==1.16.0
Pillow==9.1.1
platformdirs==2.5.1
pprintpp==0.4.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
pydocusign==2.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.12.0
pytesseract==0.3.9
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.3.1
regex==2022.6.2
reportlab==3.6.10
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sentry-sdk==1.5.8
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
svglib==1.3.0
swapper==1.3.0
tinycss2==1.1.1
tomli==2.0.1
toposort==1.7
tqdm==4.64.0
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==47
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.14.1
xhtml2pdf==0.2.8
xmltodict==0.13.0
